+++ FirebaseMLModelDownloader/Sources/ModelDownloadConditions.swift
@@ -18,7 +18,7 @@ import Foundation
 @available(
   *,
   deprecated,
-  message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+  message: "Firebase ML is deprecated and will be shut down on June 15, 2027. To host custom models, you must migrate to another solution. You can use Cloud Storage for Firebase as an alternative for hosting custom models. For more info, see https://firebase.google.com/docs/ml/migrate-to-cloud-storage"
 )
 public struct ModelDownloadConditions {
   let allowsCellularAccess: Bool