+++ FirebaseMLModelDownloader/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-- [deprecated] FirebaseMLModelDownloader has been deprecated. The service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml.
+- [deprecated] Firebase ML is deprecated and will be shut down on June 15, 2027. To host custom models, you must migrate to another solution. You can use Cloud Storage for Firebase as an alternative for hosting custom models. For more info, see https://firebase.google.com/docs/ml/migrate-to-cloud-storage.
 
 # 12.8.0
 - [fixed] Remove unused legacy telemetry along with the large swift-protobuf dependency.